+++ src/features/ServiceForm/schema.ts
@@ -19,6 +19,7 @@ const checkOutSchema = z.object({
 });
 
 const formSchema = z.object({
+    requestedCreatedAt: z.date(),
     client: z.object({
         name: z.string().min(1, "Imię jest wymagane"),
         surname: z.string().optional(),